pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.8"
dependencies = [
    "flask>=3.0.3",
    "python-dateutil>=2.9.0.post0",
]
